add posts and liked posts to user profiles
