+++ src/components/Edition/Views/Default.tsx
@@ -1,6 +1,6 @@
 import Link from 'components/Link'
 import { useUser } from 'hooks/useUser'
-import React, { useContext, useEffect, useRef } from 'react'
+import React, { useContext } from 'react'
 import { PostsContext, sortOptions } from '../Posts'
 import TableOfContents from 'components/PostLayout/TableOfContents'
 import { useLocation } from '@reach/router'
@@ -9,17 +9,16 @@ import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import isToday from 'dayjs/plugin/isToday'
 import { useLayoutData } from 'components/Layout/hooks'
-import LikeButton from '../LikeButton'
 import { CallToAction, child, container } from 'components/CallToAction'
-import Spinner from 'components/Spinner'
 import { Menu } from '@headlessui/react'
 import { IconChevronDown, IconFilter, IconX, IconSort } from '@posthog/icons'
-import { Icon, getIcon } from 'components/PostLayout/Menu'
+import { Icon } from 'components/PostLayout/Menu'
 import { navigate } from 'gatsby'
 import { postsMenu as menu } from '../../../navs/posts'
 import Intro from '../Intro'
 import Tags from '../Tags'
 import Tooltip from 'components/Tooltip'
+import PostsTable from '../PostsTable'
 dayjs.extend(relativeTime)
 dayjs.extend(isToday)
 
@@ -85,97 +84,6 @@ const CommunityBar = () => {
     )
 }
 
-const Post = ({
-    id,
-    title,
-    featuredImage,
-    date,
-    publishedAt,
-    post_category,
-    authors,
-    slug,
-    fetchMore,
-    articleView,
-}) => {
-    const containerRef = useRef()
-    const { pathname } = useLocation()
-    const category = post_category?.data?.attributes?.label
-    const active = pathname === slug
-    const breakpoints = useBreakpoint()
-    const day = dayjs(date || publishedAt)
-    const { fullWidthContent } = useLayoutData()
-
-    useEffect(() => {
-        if (active && typeof window !== 'undefined' && !breakpoints.sm) {
-            containerRef?.current?.scrollIntoView({ block: 'center', inline: 'nearest' })
-            window.scrollTo({ top: 0 })
-        }
-    }, [articleView])
-
-    const imageURL = featuredImage?.image?.data?.attributes?.url || featuredImage?.url
-    const defaultImage = post_category?.data?.attributes?.defaultImage?.data?.attributes?.url
-
-    return (
-        <li
-            ref={containerRef}
-            className={`snap-start last:pb-24 ${!articleView ? 'grid grid-cols-[28px_1fr] items-center' : ''}`}
-        >
-            {!articleView && <LikeButton postID={id} />}
-            <span className={`flex items-center ${articleView ? 'py-px' : ''}`}>
-                <Link
-                    className={`inline m-0 font-semibold border-t border-b !leading-tight line-clamp-2 text-inherit hover:text-primary dark:hover:text-primary-dark hover:text-inherit dark:text-inherit dark:hover:text-inherit hover:transition-transform flex-grow hover:bg-accent dark:hover:bg-accent-dark relative 
-                    ${
-                        active
-                            ? 'bg-accent dark:bg-accent-dark font-bold border-border dark:border-dark'
-                            : 'border-transparent hover:scale-[1.01] hover:top-[-.5px] active:top-[.5px] active:scale-[.99]'
-                    } 
-                    ${fullWidthContent ? 'p-2 ' : 'p-2 lg:px-4'}
-                    ${articleView ? 'text-[.933rem]' : 'text-base mr-1.5'}
-                    ${fullWidthContent && articleView ? 'lg:px-6' : ''} 
-                    `}
-                    to={slug}
-                >
-                    {pathname === '/posts' && category && (
-                        <p className="m-0 text-sm font-medium opacity-60 flex-shrink-0">{category}</p>
-                    )}
-                    <div
-                        className={`items-baseline ${active ? 'flex  gap-1' : ''} ${
-                            articleView ? 'flex gap-4' : 'inline'
-                        }`}
-                    >
-                        <span className="mr-1 flex-1">{title}</span>
-                        <span className={`${articleView ? 'text-right flex-0 basis-22' : 'inline-flex gap-1'}`}>
-                            <span
-                                className={`font-medium leading-none opacity-60 ${
-                                    articleView ? 'text-[.813rem]' : 'text-[.933rem]'
-                                }`}
-                            >
-                                {day.isToday() ? 'Today' : day.fromNow()}
-                            </span>
-                        </span>
-                    </div>
-                    <div className="hidden sm:w-[100px] sm:h-[85px] w-[50px] h-[50px] flex-shrink-0 bg-accent dark:bg-accent-dark rounded-sm overflow-hidden md:self-start self-center relative z-10">
-                        <span
-                            className={`text-inherit hover:text-inherit dark:text-inherit dark:hover:text-inherit flex-grow`}
-                        >
-                            {imageURL?.endsWith('.mp4') ? (
-                                <video className="object-cover w-full h-full" src={imageURL} />
-                            ) : (
-                                <img
-                                    className={`w-full h-full ${
-                                        !imageURL && defaultImage ? 'object-contain' : 'object-cover'
-                                    }`}
-                                    src={imageURL || defaultImage || '/banner.png'}
-                                />
-                            )}
-                        </span>
-                    </div>
-                </Link>
-            </span>
-        </li>
-    )
-}
-
 export const Skeleton = () => {
     return (
         <div className="flex items-center space-x-2 w-full mt-2 px-6">
@@ -185,10 +93,9 @@ export const Skeleton = () => {
     )
 }
 
-const SortDropdown = () => {
-    const { sort, setSort, articleView } = useContext(PostsContext)
+export const SortDropdown = ({ sort, setSort }) => {
     return (
-        <div className={`flex items-center ${articleView ? '' : ''}`}>
+        <div className={`flex items-center`}>
             <div className="relative">
                 <Menu>
                     <Menu.Button className="flex items-center text-sm justify-between relative pl-1.5 pt-1.5 pb-1 rounded hover:bg-light/50 hover:dark:bg-dark/50 border border-b-3 border-transparent md:hover:border-light dark:md:hover:border-dark hover:translate-y-[-1px] active:translate-y-[1px] active:transition-all">
@@ -222,7 +129,7 @@ const SortDropdown = () => {
 }
 
 export const PostFilters = ({ showTags = true, showSort = true }) => {
-    const { setRoot, setTag, tag, activeMenu, setActiveMenu, articleView } = useContext(PostsContext)
+    const { setRoot, setTag, tag, activeMenu, setActiveMenu, articleView, sort, setSort } = useContext(PostsContext)
     const { fullWidthContent } = useLayoutData()
     const breakpoints = useBreakpoint()
 
@@ -270,7 +177,7 @@ export const PostFilters = ({ showTags = true, showSort = true }) => {
                         </Menu.Items>
                     </Menu>
                 </div>
-                {showSort && <SortDropdown />}
+                {showSort && <SortDropdown sort={sort} setSort={setSort} />}
                 {showTags && activeMenu?.children?.length > 0 && (
                     <div className="flex-grow-0 flex items-center justify-center">
                         <Menu>
@@ -363,8 +270,7 @@ const Title = () => {
 }
 
 function PostsListing() {
-    const { posts, fetchMore, isValidating, isLoading, articleView, hasMore, activeMenu, tag } =
-        useContext(PostsContext)
+    const { posts, fetchMore, isValidating, isLoading, articleView, hasMore } = useContext(PostsContext)
     const breakpoints = useBreakpoint()
 
     return articleView && breakpoints.sm ? null : (
@@ -403,27 +309,14 @@ function PostsListing() {
                         articleView && !breakpoints.sm ? 'h-[85vh] overflow-auto mt-[-2px]' : ''
                     }`}
                 >
-                    {posts.map(({ id, attributes }, index) => {
-                        return <Post articleView={articleView} key={id} {...attributes} id={id} />
-                    })}
-                    {isLoading && <Skeleton />}
-                    {hasMore && (
-                        <li className="mt-4 mb-24 px-4">
-                            <button
-                                onClick={fetchMore}
-                                disabled={isLoading || isValidating}
-                                className={`${container()} w-full`}
-                            >
-                                <span className={`${child()}`}>
-                                    {isLoading || isValidating ? (
-                                        <Spinner className="!w-6 !h-6 mx-auto text-white" />
-                                    ) : (
-                                        'Load more'
-                                    )}
-                                </span>
-                            </button>
-                        </li>
-                    )}
+                    <PostsTable
+                        posts={posts}
+                        fetchMore={fetchMore}
+                        isValidating={isValidating}
+                        isLoading={isLoading}
+                        hasMore={hasMore}
+                        articleView={articleView}
+                    />
                 </ul>
             </div>
         </div>
